requirements.txt
django==2.2.7
djangorestframework==3.10.3

requests==2.22.0

pillow==6.2.1
